scripts/requirements.txt
Flask
SQLAlchemy
Faker
python-dotenv
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Flask-SocketIO==5.5.1
psycopg2-binary==2.9.10